Updated Lobby water and radiance cache disabled
Updated the water to be a translucent shader and used a material under as the drive for color.
Client mentioned to make the water less reflective. To my understanding what they meant is that the water wasnt "see-through" enough. Hopefully this makes it more "realistic".

Also added command execution r.Lumen.TranslucencyReflections.RadianceCache 0 to remove unwanted reflection artifacts